refactor: change path from shared/common/types/AppTypes to shared/common/types
AppTypes folder has been  deleted
