+++ src/modules/auth/ForgotPassword/types/ForgotPassword.types.ts
@@ -1,5 +1,5 @@
 import { FieldErrors, UseFormRegister } from "react-hook-form";
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { FunctionReturnsVoid } from "shared/common/types";
 
 export interface IForgotPasswordForm {
 	email: string;
