+++ src/modules/dashboard/Board/components/Dialogs/DialogDeleteOneBoard/types/DialogDeleteOneBoard.types.ts
@@ -1,4 +1,4 @@
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { FunctionReturnsVoid } from "shared/common/types";
 
 export interface IDialogDeleteOneBoardView {
 	isOpenDialogDeleteOneBoard: boolean;
