+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/types/DialogEditBoard.types.ts
@@ -3,7 +3,7 @@ import {
 	UseFormHandleSubmit,
 	UseFormRegister,
 } from "react-hook-form";
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { FunctionReturnsVoid } from "shared/common/types";
 
 export interface IDialogEditBoardForm {
 	title: string;
