+++ src/modules/dashboard/Board/shared/hook/useDialogBoard/types/useDialogBoard.types.ts
@@ -1,4 +1,4 @@
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { FunctionReturnsVoid } from "shared/common/types";
 
 export interface IUseDialogBoard {
 	openDialogEditBoard: (closeMenu: FunctionReturnsVoid) => void;
