+++ src/modules/error/Page404/types/Page404.component.ts
@@ -1,4 +1,4 @@
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { FunctionReturnsVoid } from "shared/common/types";
 
 export interface IPage404View {
 	backPage: FunctionReturnsVoid;
