+++ src/modules/notes/Note/Context/types/Note.context.types.ts
@@ -1,5 +1,5 @@
 import React from "react";
-import { ViewMode } from "shared/common/types/AppTypes";
+import { ViewMode } from "shared/common/types";
 
 export interface IContextProviderNote {
 	children: React.ReactNode;
