+++ src/modules/notes/Note/components/Dialogs/DialogDeleteOneNote/types/DialogDeleteOneNote.types.ts
@@ -1,6 +1,6 @@
 import { UseMutateFunction } from "@tanstack/react-query";
 import { AxiosResponse } from "axios";
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { FunctionReturnsVoid } from "shared/common/types";
 import { IFetchResponseDefault } from "shared/common/types/Fetch";
 
 export interface IDialogDeleteOneNoteView {
