+++ src/modules/notes/Note/shared/hook/useDialogNote/types/useDialogNote.types.ts
@@ -1,4 +1,4 @@
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { FunctionReturnsVoid } from "shared/common/types";
 
 export interface IUseDialogNote {
 	openDialogDeleteOneNote: (noteID: string) => void;
