+++ src/modules/texts/Text/components/HeaderText/types/HeaderText.types.ts
@@ -1,4 +1,4 @@
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { FunctionReturnsVoid } from "shared/common/types";
 
 export interface IHeaderText {
 	count: number | undefined;
