+++ src/modules/texts/Text/shared/hooks/useDialogText/types/useDialogText.types.ts
@@ -1,4 +1,4 @@
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { FunctionReturnsVoid } from "shared/common/types";
 
 export interface IUseDialogText {
 	openDialogDeleteOneText: (textID: string) => void;
