+++ src/modules/texts/TextEdit/components/Markdown/components/HeaderText/types/HeaderText.types.ts
@@ -1,5 +1,5 @@
 import { ChangeEvent } from "react";
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { FunctionReturnsVoid } from "shared/common/types";
 
 export interface IHeaderText {
 	handleClickSaveText: FunctionReturnsVoid;
