+++ src/shared/common/Context/ThemeContext/types/ThemeContext.types.ts
@@ -1,4 +1,4 @@
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { FunctionReturnsVoid } from "shared/common/types";
 
 export type ThemeName = "dark" | "light";
