+++ src/shared/components/Drawer/components/Aside/components/LinkTreeItem/utils/define-url-path/__test__/define-url-path.test.ts
@@ -1,4 +1,4 @@
-import { TypeBoard } from "shared/common/types/AppTypes";
+import { TypeBoard } from "shared/common/types";
 import {
 	defineURLPath,
 	pathMapping,
