+++ src/shared/components/Drawer/components/DialogSearchAll/components/TableDialogSearchAll/utils/generateQueryParams/types/generateQueryParams.types.ts
@@ -1,4 +1,4 @@
-import { TypeBoard } from "shared/common/types/AppTypes";
+import { TypeBoard } from "shared/common/types";
 
 export interface IGenerateQueryParams {
 	id: string;
