+++ src/shared/components/Drawer/components/DialogSearchAll/types/DialogSearchAll.types.ts
@@ -5,8 +5,8 @@ import {
 	UseFormRegister,
 } from "react-hook-form";
 import { IDialogSearchAllState } from "shared/common/store/DialogSearchAll/types/DialogSearchAll.types";
-import { IPaginationModel } from "shared/common/types/AppTypes";
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { IPaginationModel } from "shared/common/types";
+import { FunctionReturnsVoid } from "shared/common/types";
 import { IFetchGetSearchAllResponse } from "shared/components/Drawer/components/DialogSearchAll/services/types";
 
 export interface IDialogSearchAllForm {
