+++ src/shared/components/Drawer/components/Header/types/Header.component.ts
@@ -1,4 +1,4 @@
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { FunctionReturnsVoid } from "shared/common/types";
 
 export interface IHeaderView {
 	isOpen: boolean;
