+++ src/shared/components/Layout/Snackbar/types/Snackbar.types.ts
@@ -1,5 +1,5 @@
 import { ISnackbarProps } from "shared/common/hook/useSnackbar/types";
-import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
+import { FunctionReturnsVoid } from "shared/common/types";
 
 export interface ISnackbarView {
 	snackbarProps: ISnackbarProps;
